requirements.txt
absl-py==0.9.0
astor==0.8.1
autopep8==1.5
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
CNFgen==0.8.4.1
cycler==0.10.0
decorator==4.4.2
dgl-cu101==0.5.0
dgl-cu110==0.5.3
entrypoints==0.3
flake8==3.7.9
future==0.18.2
gast==0.2.2
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna==2.9
joblib==0.15.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lmdb==0.98
Markdown==3.2.1
matplotlib==3.2.1
mccabe==0.6.1
networkx==2.4
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.2.0
pandas==1.0.2
Pillow==7.0.0
progress==1.5
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pydot==1.4.1
pyflakes==2.1.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytorch-swats==0.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
typing-extensions==3.7.4.2
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.12.1
